+++ plonk/analysis/profile.py
@@ -29,7 +29,7 @@ from .._logging import logger
 from .._units import Quantity
 from .._units import units as plonk_units
 from ..utils.math import average
-from ..utils.utils import is_documented_by, pretty_array_name
+from ..utils.strings import is_documented_by, pretty_array_name
 from ._profiles import extra_profiles
 
 if TYPE_CHECKING: