+++ plonk/snap/__init__.py
@@ -6,14 +6,11 @@ snapshot file.
 
 from .readers import load_snap
 from .snap import Sinks, Snap, SnapLike, SubSnap
-from .utils import dust_array_names, gravitational_constant_in_code_units
 
 __all__ = [
     'Sinks',
     'Snap',
     'SnapLike',
     'SubSnap',
-    'dust_array_names',
-    'gravitational_constant_in_code_units',
     'load_snap',
 ]