+++ plonk/visualize/__init__.py
@@ -5,7 +5,6 @@ simulations using kernel density estimation based interpolation.
 """
 
 from .animation import animation, animation_particles, animation_profiles
-from .functions import get_extent_from_percentile, plot_smoothing_length
 from .interpolation import interpolate
 from .simulation import visualize_sim
 from .visualization import image, plot, vector
@@ -14,9 +13,7 @@ __all__ = [
     'animation',
     'animation_particles',
     'animation_profiles',
-    'get_extent_from_percentile',
     'image',
-    'plot_smoothing_length',
     'interpolate',
     'plot',
     'vector',