+++ tests/test_visualization.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 
 import plonk
-from plonk import visualize
+from plonk.utils import visualize
 
 AU = plonk.units('au')
 TEST_FILE = Path(__file__).parent / 'stubdata/phantom_00000.h5'